PILOT-82 | use flask-classy, webargs, marshmallow for REST development
