+++ app/api_1_0/errors.py
@@ -1,18 +1,13 @@
-# -*- coding: utf-8 -*-
-
-"""api v1.0 error handlers"""
-
-from flask import request, jsonify, render_template
+from flask import request, render_template
 from sqlalchemy.exc import SQLAlchemyError
 from jwt.exceptions import InvalidTokenError
 
 from ..api.errors import api_exception_handler
 from ..exceptions import ApplicationException, NotFoundException, BadRequestException
+
 from . import api_bp
 
-
 api_bp.errorhandler(ApplicationException)(api_exception_handler)
-
 
 @api_bp.errorhandler(ValueError)
 def value_error_handler(ex):
@@ -38,7 +33,7 @@ def invalid_token_error_handler(ex):
     return api_exception_handler(ex)
 
 
-@api_bp.app_errorhandler(404)
+@api_bp.errorhandler(404)
 def error_handler(ex):
     """Register error handler for 404"""
     if request.path.startswith('/api/'):  # FIXME(hoatle): hard-code
