+++ app/blueprints.py
@@ -4,7 +4,6 @@
 
 from .main import main_bp
 from .api_1_0 import api_bp as api_1_0_bp
-from .api_2_0 import api_bp as api_2_0_bp
 
 __all__ = ['register_blueprints']
 
@@ -13,4 +12,3 @@ def register_blueprints(app):
     """register blueprints"""
     app.register_blueprint(main_bp)
     app.register_blueprint(api_1_0_bp)
-    app.register_blueprint(api_2_0_bp)
