+++ tests/integration/api_1_0/test_roles.py
@@ -3,9 +3,9 @@ import json
 from tests.integration import IntegrationTestCase
 
 
-class RoleListAPITestCase(IntegrationTestCase):
+class RoleResourceTestCase(IntegrationTestCase):
 
-    def test_get_invalid_auth(self):
+    def test_index_invalid_auth(self):
         rv = self.client.get('/api/v1.0/roles')
         self.assertEqual(rv.status_code, 401)
         resp = json.loads(rv.data)
@@ -13,13 +13,13 @@ class RoleListAPITestCase(IntegrationTestCase):
         self.assertEqual(error.get('message'), 'Authorization Required')
         self.assertEqual(error.get('description'), 'Authorization header was missing')
 
-    def test_get_invalid_perm(self):
+    def test_index_invalid_perm(self):
         pass
 
-    def test_get_valid_perm(self):
+    def test_index_valid_perm(self):
         pass
 
-    def test_post_invalid_auth(self):
+    def test_create_invalid_auth(self):
         rv = self.client.post('/api/v1.0/roles')
         self.assertEqual(rv.status_code, 401)
         resp = json.loads(rv.data)
@@ -27,16 +27,14 @@ class RoleListAPITestCase(IntegrationTestCase):
         self.assertEqual(error.get('message'), 'Authorization Required')
         self.assertEqual(error.get('description'), 'Authorization header was missing')
 
-    def test_post_invalid_perm(self):
+    def test_create_invalid_perm(self):
         pass
 
-    def test_post_valid_perm(self):
+    def test_create_valid_perm(self):
         pass
 
 
-class RoleAPITestCase(IntegrationTestCase):
-
-    def test_get_invalid_auth(self):
+    def test_show_invalid_auth(self):
         rv = self.client.get('/api/v1.0/roles/1')
         self.assertEqual(rv.status_code, 401)
         resp = json.loads(rv.data)
@@ -44,13 +42,13 @@ class RoleAPITestCase(IntegrationTestCase):
         self.assertEqual(error.get('message'), 'Authorization Required')
         self.assertEqual(error.get('description'), 'Authorization header was missing')
 
-    def test_get_invalid_perm(self):
+    def test_show_invalid_perm(self):
         pass
 
-    def test_get_valid_perm(self):
+    def test_show_valid_perm(self):
         pass
 
-    def test_put_invalid_auth(self):
+    def test_update_invalid_auth(self):
         rv = self.client.put('/api/v1.0/roles/1')
         self.assertEqual(rv.status_code, 401)
         resp = json.loads(rv.data)
@@ -58,13 +56,13 @@ class RoleAPITestCase(IntegrationTestCase):
         self.assertEqual(error.get('message'), 'Authorization Required')
         self.assertEqual(error.get('description'), 'Authorization header was missing')
 
-    def test_put_invalid_perm(self):
+    def test_update_invalid_perm(self):
         pass
 
-    def test_put_valid_perm(self):
+    def test_update_valid_perm(self):
         pass
 
-    def test_delete_invalid_auth(self):
+    def test_destroy_invalid_auth(self):
         rv = self.client.delete('/api/v1.0/roles/1')
         self.assertEqual(rv.status_code, 401)
         resp = json.loads(rv.data)
@@ -72,8 +70,8 @@ class RoleAPITestCase(IntegrationTestCase):
         self.assertEqual(error.get('message'), 'Authorization Required')
         self.assertEqual(error.get('description'), 'Authorization header was missing')
 
-    def test_delete_invalid_perm(self):
+    def test_destroy_invalid_perm(self):
         pass
 
-    def test_delete_valid_perm(self):
+    def test_destroy_valid_perm(self):
         pass
