+++ tests/unit/app/api/test_base.py
@@ -244,275 +244,15 @@ class BaseTestCase(CurrentAppMockMixin, UnitTestCase):
             self.assertEqual(SA_OPS_MAPPER.get(key), value, '{} should match {}'.format(key, value))
 
 
-class RequestParserResourceMixinTestCase(CurrentAppMockMixin, UnitTestCase):
-    def test_init(self):
-        from app.api.base import RequestParserResourceMixin
-
-        req_parser_resource = RequestParserResourceMixin()
-        req_parses = req_parser_resource.req_parses
-        self.assertEqual(len(req_parses), 3, 'len(req_parses) should be 3')
-        self.assertTrue('common' in req_parses, "'common' should be in req_parses")
-        self.assertTrue('get' in req_parses, "'get' should be in req_parses")
-        self.assertTrue('post' in req_parses, "'post' should be in req_parses")
-
-    def test_add_argument(self):
-        from app.api.base import RequestParserResourceMixin
-
-        req_parser_resource = RequestParserResourceMixin()
-        req_parser_resource.add_argument('get', 'name', str)
-        self.assertIsNotNone(req_parser_resource.req_parses['get'],
-                             "req_parser_resource.req_parses['get'] should not be None")
-
-        req_parser_resource.add_argument('put', 'email', str)
-        self.assertIsNotNone(req_parser_resource.req_parses['put'],
-                             "req_parser_resource.req_parses['put'] should not be None")
-
-        self.assertEqual(len(req_parser_resource.req_parses), 4,
-                         'len(req_parser_resource.req_parses) should be 4')
-
-    @patch('app.api.base.RequestParser')
-    @patch('app.api.base.request')
-    def test_parse_arguments_common_args(self, mock_request, mock_request_parser):
-        from app.api.base import RequestParserResourceMixin
-
-        mock_request.method.lower.return_value = 'put'
-
-        req_parse_resource = RequestParserResourceMixin()
-
-        common_req_parser = req_parse_resource.req_parses['common']
-        common_req_parser.parse_args.return_value = {
-            'lang': None
-        }
-
-        args = req_parse_resource.parse_arguments()
-
-        self.assertTrue(mock_request_parser.called)
-        self.assertEqual(args, {}, 'args should be {}')
-
-    @patch('app.api.base.RequestParser')
-    @patch('app.api.base.request')
-    def test_parse_arguments(self, mock_request, mock_request_parser):
-        from app.api.base import RequestParserResourceMixin
-
-        mock_request.method.lower.return_value = 'get'
-
-        req_parse_resource = RequestParserResourceMixin()
-
-        common_req_parser = req_parse_resource.req_parses['common']
-        common_req_parser.parse_args.side_effect = [
-            {
-                'lang': None
-            },
-            {
-                'name': 'api',
-                'status': 'done'
-            }
-        ]
-
-        args = req_parse_resource.parse_arguments()
-
-        self.assertTrue(mock_request_parser.called)
-
-        self.assertEqual(args, {
-            'name': 'api',
-            'status': 'done'
-        })
-
-    @patch('app.api.base.RequestParser')
-    @patch('app.api.base.request')
-    def test_parse_arguments_filters(self, mock_request, mock_request_parser):
-        from app.api.base import RequestParserResourceMixin
-
-        mock_request.method.lower.return_value = 'get'
-
-        req_parse_resource = RequestParserResourceMixin()
-
-        common_req_parser = req_parse_resource.req_parses['common']
-        common_req_parser.parse_args.side_effect = [
-            {
-                'lang': 'en'
-            },
-            {
-                'name': 'api',
-                'a__eq': 1,
-                'b__ne': 2,
-                'c__lt': 3,
-                'd__le': 4,
-                'e__gt': 5,
-                'f__ge': 6,
-                'g__lk': '%a',
-                'h__nl': '%b%',
-                'i__in': 'c,d,e',
-                'j__ni': 'f,g,h',
-                'k__ct': 'abc',
-                'l__mc': 'def',
-                'm___eq': 'no'
-            },
-            {
-                'lang': 'en'
-            },
-            {
-                'n__eq': 7
-            }
-        ]
-
-        args = req_parse_resource.parse_arguments()
-
-        self.assertEqual(len(args), 4)
-        self.assertEqual(args['lang'], 'en')
-        self.assertEqual(args['name'], 'api')
-        self.assertEqual(args['m___eq'], 'no')
-        self.assertTrue('filters' in args)
-
-        filters = args.get('filters')
-
-        self.assertEqual(len(filters), 12)
-
-        check_list = [
-            {
-                'key': 'a',
-                'op': 'eq',
-                'value': 1
-            },
-            {
-                'key': 'b',
-                'op': 'ne',
-                'value': 2
-            },
-            {
-                'key': 'c',
-                'op': 'lt',
-                'value': 3
-            },
-            {
-                'key': 'd',
-                'op': 'le',
-                'value': 4
-            },
-            {
-                'key': 'e',
-                'op': 'gt',
-                'value': 5
-            },
-            {
-                'key': 'f',
-                'op': 'ge',
-                'value': 6
-            },
-            {
-                'key': 'g',
-                'op': 'like',
-                'value': '%a'
-            },
-            {
-                'key': 'h',
-                'op': 'notlike',
-                'value': '%b%'
-            },
-            {
-                'key': 'i',
-                'op': 'in_',
-                'value': ['c', 'd', 'e']
-            },
-            {
-                'key': 'j',
-                'op': 'notin_',
-                'value': ['f', 'g', 'h']
-            },
-            {
-                'key': 'k',
-                'op': 'contains',
-                'value': 'abc'
-            },
-            {
-                'key': 'l',
-                'op': 'match',
-                'value': 'def'
-            }
-        ]
-
-        for item in check_list:
-            self.assertTrue(item in filters, '{} should be in {}'.format(item, filters))
-
-        # bug fix for filters on another request parser
-        args = req_parse_resource.parse_arguments()
-
-        self.assertEqual(len(args), 2)
-        self.assertEqual(args['lang'], 'en')
-        self.assertTrue('filters' in args)
-
-        filters = args.get('filters')
-
-        self.assertEqual(len(filters), 1)
-
-        a_filter = {
-            'key': 'n',
-            'op': 'eq',
-            'value': 7
-        }
-        self.assertTrue(a_filter in filters)
-
-
-class BaseResourceTestCase(CurrentAppMockMixin, UnitTestCase):
-    def test_class(self):
-        from app.api.base import BaseResource, RequestParserResourceMixin
-
-        self.assertTrue(issubclass(BaseResource, Resource))
-        self.assertTrue(issubclass(BaseResource, RequestParserResourceMixin))
-        self.assertEqual(BaseResource.action_decorators, {},
-                         'BaseResource.action_decorators should be {}')
-
-    @patch('app.api.base.Resource.dispatch_request')
-    @patch('app.api.base.request')
-    def test_dispatch_request(self, mock_request, mock_parent_dispatch_request):
-        from app.api.base import BaseResource
-
-        mock_request.method.lower.return_value = 'get'
-        mock_parent_dispatch_request.return_value = 'test_called'
-
-        base_resource = BaseResource()
-        rv = base_resource.dispatch_request(1, lang='en')
-
-        self.assertEqual(rv, 'test_called')
-        mock_request.method.lower.assert_called_once_with()
-        mock_parent_dispatch_request.assert_called_once_with(1, lang='en')
-
-    @patch('app.api.base.Resource.dispatch_request')
-    @patch('app.api.base.request')
-    def test_action_decorators(self, mock_request, mock_parent_dispatch_request):
-        from functools import wraps
-        from app.api.base import BaseResource
-
-        mock = MagicMock()
-
-        def test_decorator(fn):
-            @wraps(fn)
-            def decorated(*args, **kwargs):
-                mock()
-                return fn(*args, **kwargs)
-
-            return decorated
-
-        base_resource = BaseResource()
-        base_resource.action_decorators = {
-            'get': [test_decorator]
-        }
-        mock_request.method.lower.return_value = 'get'
-        mock_parent_dispatch_request.__name__ = 'foo'
-        mock_parent_dispatch_request.return_value = 'hi'
-
-        rv = base_resource.dispatch_request('users', active=True)
-        self.assertEqual(rv, 'hi')
-        self.assertTrue(mock.called, 'mock.called should be True')
-        mock_request.method.lower.assert_called_once_with()
-        mock_parent_dispatch_request.assert_called_once_with('users', active=True)
-
+class ResourceTestCase(CurrentAppMockMixin, UnitTestCase):
+    def test_something(self):
+        pass
 
 class TokenRequiredResourceTestCase(CurrentAppMockMixin, UnitTestCase):
     def test_class(self):
-        from app.api.base import TokenRequiredResource, BaseResource
+        from app.api.base import TokenRequiredResource, Resource
 
-        self.assertTrue(issubclass(TokenRequiredResource, BaseResource))
+        self.assertTrue(issubclass(TokenRequiredResource, Resource))
 
     def test_decorators(self):
         from app.api.base import TokenRequiredResource
@@ -524,9 +264,9 @@ class TokenRequiredResourceTestCase(CurrentAppMockMixin, UnitTestCase):
 
 class AdminRoleRequiredResourceTestCase(CurrentAppMockMixin, UnitTestCase):
     def test_class(self):
-        from app.api.base import AdminRoleRequiredResource, BaseResource
+        from app.api.base import AdminRoleRequiredResource, Resource
 
-        self.assertTrue(issubclass(AdminRoleRequiredResource, BaseResource))
+        self.assertTrue(issubclass(AdminRoleRequiredResource, Resource))
 
     def test_decorators(self):
         from app.api.base import AdminRoleRequiredResource
