--- tests/unit/app/api_1_0/test_fields.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-"""tests for api_1_0.fields"""
