+++ tests/unit/app/api_1_0/test_users.py
@@ -1,3 +1,11 @@
 # -*- coding: utf-8 -*-
 
 """tests for api_1_0.users"""
+
+from tests.unit import UnitTestCase
+
+
+class UserResourceTestCase(UnitTestCase):
+
+    def test_something(self):
+        pass
